+++ src/core/renderer/tiles/optimizedTraverseFunctions.js
@@ -28,7 +28,10 @@ function isProcessed( tile ) {
 // Checks whether all children have been processed and are ready to traverse
 function areChildrenProcessed( tile ) {
 
-	return tile.internal.childrenProcessed === tile.children.length && ( ! tile.internal.hasUnrenderableContent || isDownloadFinished( tile.internal.loadingState ) );
+	// all children are processed at once
+	const childrenReady = tile.children.length === 0 || Boolean( tile.children[ 0 ].internal );
+	const contentReady = ! tile.internal.hasUnrenderableContent || isDownloadFinished( tile.internal.loadingState );
+	return childrenReady && contentReady;
 
 }