+++ src/core/renderer/tiles/traverseFunctions.js
@@ -33,7 +33,8 @@ function isProcessed( tile ) {
 
 function areChildrenProcessed( tile ) {
 
-	return tile.internal.childrenProcessed === tile.children.length;
+	// all children are processed at once
+	return tile.children.length === 0 || Boolean( tile.children[ 0 ].internal );
 
 }